Speed Improvements
Now uses a spatial index for improved performance.  Also, now campares
the actual area of the intersection between two buildings with the area
of each building, vs only using the areas of their bounding boxes.  This
reduces the number of false positives (buildings that were selected when
they only overlapped a small amount).